Simplify initial background pattern generation
Originally we had the idea of keeping a seed in storage and consistently
generating the CSS background-image based on that seed. However, we
ended up going for an approach where the full (un)trusted CSS is kept in
storage instead (also because this gives greater flexibility for the
user in modifying more styles as they wish). This allowed us to simplify
the color generation process, getting rid of seed and hashes, and just
generating a random value.